requirements.txt
fastapi
openai
pydantic
uvicorn
pandas
pygithub